changes to make renderless component up to date with new project structure, adding updates to tests and temporary modifications to stories before major refactor on stories
