+++ packages/components/organisms/f-content-cards/src/components/cardTemplates/homePromotionCard/HomePromotionCard1.vue
@@ -7,18 +7,10 @@
             :data-test-id="containerTestId"
             :class="['l-container', 'c-contentCards-homePromotionCard1-container', $style['c-contentCards-homePromotionCard1-container']]"
             :style="{ maxWidth: `${containerMaxWidth}px` }">
-            <div :class="['c-contentCards-homePromotionCard1-iconPane', $style['c-contentCards-homePromotionCard1-iconPane']]">
+            <div :class="['c-contentCards-homePromotionCard1-icon', $style['c-contentCards-homePromotionCard1-icon']]">
                 <img
-                    :class="[$style['c-contentCards-homePromotionCard1-icon']]"
                     :src="icon"
                     alt="">
-                <h3
-                    :class="[$style['c-contentCards-homePromotionCard1-subtitle'], {
-                        [$style['c-contentCards-homePromotionCard1-subtitle--light']]: isLightSubtitle
-                    }]"
-                >
-                    {{ subtitle }}
-                </h3>
             </div>
             <div :class="['c-contentCards-homePromotionCard1-innerCard', $style['c-contentCards-homePromotionCard1-innerCard']]">
                 <home-promotion-card2 :card="card" />
@@ -28,7 +20,6 @@
 </template>
 
 <script>
-import Color from 'color';
 import HomePromotionCard2 from './HomePromotionCard2.vue';
 
 export default {
@@ -60,8 +51,7 @@ export default {
             icon,
             title,
             url,
-            description,
-            subtitle
+            description
         } = this.card;
 
         return {
@@ -74,8 +64,7 @@ export default {
             type,
             icon,
             url,
-            description,
-            subtitle
+            description
         };
     },
     computed: {
@@ -85,20 +74,6 @@ export default {
 
         containerTestId () {
             return this.testId ? `${this.testId}--container` : false;
-        },
-
-        /**
-         * If background colour is set *and* dark, then use a light text colour for the subtitle for A11y
-         * @return {boolean}
-         */
-        isLightSubtitle () {
-            try {
-                return this.backgroundColor
-                    ? (new Color(this.backgroundColor)).isDark()
-                    : false;
-            } catch {
-                return false;
-            }
         }
     }
 };
@@ -120,36 +95,23 @@ export default {
         margin: 0 auto;
     }
 
-    .c-contentCards-homePromotionCard1-iconPane {
+    .c-contentCards-homePromotionCard1-icon {
         display: flex;
-        flex-direction: column;
         align-items: center;
         justify-content: center;
         width: 100%;
         margin-bottom: spacing(x3);
+
+        img {
+            max-height: 65px;
+            max-width: 100%;
+        }
 
         @include media('>mid') {
             width: 50%;
             margin-bottom: 0;
         }
     }
-
-    .c-contentCards-homePromotionCard1-icon {
-        max-height: 88px;
-        max-width: 100%;
-    }
-
-    .c-contentCards-homePromotionCard1-subtitle {
-        display: none;
-
-        @include media('>mid') {
-            display: unset;
-        }
-    }
-
-        .c-contentCards-homePromotionCard1-subtitle--light {
-            color: $white;
-        }
 
     .c-contentCards-homePromotionCard1-innerCard {
         width: 100%;
@@ -160,20 +122,6 @@ export default {
             width: 50%;
             padding-left: 0;
             padding-right: spacing(x4);
-
-            :global(.c-contentCards-homePromotionCard2) {
-                padding-left: spacing(x5);
-            }
-
-            :global(.c-contentCards-homePromotionCard2-title) {
-                @include font-size(heading-m);
-                margin-bottom: spacing();
-            }
-
-            :global(.c-contentCards-homePromotionCard2-text) {
-                @include font-size(subheading-s);
-                margin-top: spacing();
-            }
         }
     }
 </style>
