+++ packages/components/organisms/f-content-cards/src/components/cardTemplates/index.js
@@ -4,9 +4,10 @@ import PostOrderCard from './PostOrderCard.vue';
 import SkeletonLoader from './SkeletonLoader.vue';
 import TermsAndConditionsCard from './TermsAndConditionsCard.vue';
 import VoucherCard from './voucherCard/VoucherCard.vue';
+import GroupHeaderCard from './GroupHeaderCard.vue';
 import { HomePromotionCard1, HomePromotionCard2 } from './homePromotionCard';
 
-export default {
+export {
     FirstTimeCustomerCard,
     PromotionCard,
     PostOrderCard,
@@ -14,5 +15,6 @@ export default {
     TermsAndConditionsCard,
     VoucherCard,
     HomePromotionCard1,
-    HomePromotionCard2
+    HomePromotionCard2,
+    GroupHeaderCard
 };
